Create `Layout`-component
