+++ apps/website-25/src/components/Layout.tsx
@@ -1,0 +1,30 @@
+import { EXTERNAL_LINK_PROPS, Footer, Nav } from '@bluedot/ui';
+
+// TODO: 01/27 add routing to courses when AISafetyFundamentals course is integrated, i.e.'/courses/intro-transformative-ai
+const courses = [
+  { title: 'Intro to Transformative AI', href: 'https://aisafetyfundamentals.com/intro-to-tai/' },
+  { title: 'AI Alignment Fast-Track', href: 'https://aisafetyfundamentals.com/alignment-fast-track/' },
+  { title: 'AI Alignment In-Depth', href: 'https://aisafetyfundamentals.com/alignment/' },
+  { title: 'AI Governance Fast-Track', href: 'https://aisafetyfundamentals.com/governance-fast-track/' },
+  { title: 'Economics of Transformative AI Fast-Track', href: 'https://aisafetyfundamentals.com/economics-of-tai-fast-track/', isNew: true },
+];
+
+const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
+  return (
+    <>
+      <Nav
+        logo="/images/logo/BlueDot_Impact_Logo.svg"
+        courses={courses}
+      >
+        <a href="https://donate.stripe.com/5kA3fpgjpdJv6o89AA" {...EXTERNAL_LINK_PROPS}>Support us</a>
+        <a href="/about">About</a>
+        <a href="/careers">Join us</a>
+        <a href="https://bluedot.org/blog/" {...EXTERNAL_LINK_PROPS}>Blog</a>
+      </Nav>
+      <main className="bluedot-base">{children}</main>
+      <Footer logo="/images/logo/BlueDot_Impact_Logo_White.svg" />
+    </>
+  );
+};
+
+export default Layout;
